CIV-11541 Update state event added
