+++ src/main/java/uk/gov/hmcts/reform/civil/callback/CaseEvent.java
@@ -303,6 +303,7 @@ public enum CaseEvent {
     CREATE_SERVICE_REQUEST_API_HMC(CAMUNDA),
     UPDATE_PARTIES_NOTIFIED_HMC(CAMUNDA),
     UPDATE_CASE_PROGRESS_HMC(CAMUNDA),
+    UPDATE_CLAIMANT_INTENTION_CLAIM_STATE(CAMUNDA),
     NOTIFY_APPLICANT1_CLAIM_SUBMITTED(CAMUNDA),
     NOTIFY_DEFENDANT_CLAIMANT_SETTLE_THE_CLAIM(CAMUNDA),
     SEND_CVP_JOIN_LINK(USER),
